AIOPS/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.12.5
GitPython==3.1.46
python-dotenv==1.2.2
pytest==9.0.2
